final_project/back-end/requirements.txt
ffmpy3==0.2.4
Flask==2.0.3
numpy==1.21.5
requests==2.27.1
websocket_client==1.3.2
